Added working FeatureDetector and changed the min SDK
FeatureDetector is now working in real time. I have also added the
cleanup onDestroy and the onPause like it is from OpenCV samples.

Change of minSDK and compileSDK was required so that I could test it on
my phone. It will stay as it enables a bit more flexible debugging than
Emulator.

SurfaceView below the camera is intended to draw the path/map of your
movement.