Added SurfaceView which will serve us as a relative map drawing
Beside few minor fancy fixes like back button working on main activity
to nicely finish it and implementing temporary random generator to
generate x and y points so that we can test the map drawing, the general
goal was to create a drawable map. The map with blue background already
appears, runnable inside the drawMap runs and gets points from main
application (even thou they are through global variables atm), but I
need to figure out why it is also not displayed on the SurfaceView.