RigidTransformation manages to trace the camera movement and draw a map
After some experimenting I just took the simplest method I could find
to confirm that drawing a map is possible. There are left-right up-down
issues on some camera that need to be solved but this is fully
functional code for drawing on SurfaceView with Path and to update the
UI in separate thread.

In near future it might be good to move all other frame processing in
camera in other thread. Also refactoring so we will be able to pick
method from menu is needed.